Split up by lines in server_scpi_io()

+++ src/parser.h
@@ -20,7 +20,7 @@ extern void parser_input(
 
 extern int parser_init(struct info *info);
 extern int parser_done(struct info *info);
-extern int parser_send(struct info *info, char *line, size_t len);
+extern int parser_send(struct info *info, char *line, int len);
 extern int parser_num(const char *s, struct scpi_type *val, int token);
 
 #endif /* PARSER_H_ */
